fix: allow matrix input to optimal filter
In optimal filtering with know correlation,
an optimal diagonal filter is obtained with
given correlation matrices and the input
is multiplied with this filter.

Matrix input (joint time-vertex signal) can be
allowed since it corresponds to finding an optimal
diagonal filter with given matrices and applying
it to every column individually. This is also
what we want in the simplified optimal filtering
task.

Refs: #6